Renamed parser files from linebased and moved them up
