+++ lex-parser/src/lex/parsing.rs
@@ -22,8 +22,10 @@
 
 // Parser implementations
 pub mod common;
+pub mod engine;
 pub mod ir;
-pub mod linebased;
+pub mod parser;
+pub mod tree_builder;
 // Temporary keepers for future analyzer experiments remain available via the
 // AnalyzerConfig infrastructure in crate::lex::parsing::pipeline.
